# api-documentation.yaml
openapi: 3.0.3
info:
  title: AI-HR Recruitment System API
  description: |
    Complete AI-driven recruitment system with resume filtering, AI interview chat, and candidate scoring functionality.

    ## System Workflow
    1. **User Management**: Candidate registration/login, HR login and create other HR accounts
    2. **Job Management**: HR creates and closes job positions
    3. **Application Process**: Candidates apply for positions and submit resumes
    4. **AI Assessment**: AI filters resumes and scores, chats with candidates and scores again
    5. **HR Review**: HR views job applicants, scores, and resumes

    ## Authentication
    Uses RSA-signed JWT Bearer Token for authentication.

  version: 2.0.0
  contact:
    name: AI-HR Team
    email: yifei.liu@tum.de
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development environment

# Security configuration
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RSA-signed JWT token

  # Data model definitions
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        message:
          type: string
          description: Response message
        data:
          oneOf:
            - type: object
            - type: array
            - type: "null"
          description: Response data
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        code:
          type: integer
          description: HTTP status code
      required:
        - success
        - message
        - timestamp
        - code

    # Error response
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error description"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        timestamp:
          type: string
          format: date-time
        code:
          type: integer

    # Pagination information
    PageInfo:
      type: object
      properties:
        page:
          type: integer
          minimum: 0
          description: Current page number (0-based)
        size:
          type: integer
          minimum: 1
          maximum: 100
          description: Number of items per page
        total:
          type: integer
          description: Total number of items
        totalPages:
          type: integer
          description: Total number of pages

    # User-related models
    UserDto:
      type: object
      properties:
        userID:
          type: string
          format: uuid
        fullName:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
          maxLength: 255
        role:
          type: string
          enum: [CANDIDATE, HR]
        creationTimestamp:
          type: string
          format: date-time
      required:
        - userID
        - fullName
        - email
        - role

    # Authentication-related models
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - email
        - password
    RegisterRequest:
      type: object
      properties:
        fullName:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 6
      required:
        - fullName
        - email
        - password
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        user:
          $ref: "#/components/schemas/UserDto"
        expiresIn:
          type: integer
          description: Access token expiration time (seconds)

    # Job position-related models
    JobDto:
      type: object
      properties:
        jobID:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 255
        description:
          type: string
        requirements:
          type: string
          description: Job requirements
        status:
          type: string
          enum: [OPEN, CLOSED, DRAFT]
        creationTimestamp:
          type: string
          format: date-time
        closingDate:
          type: string
          format: date
        lastModifiedTimestamp:
          type: string
          format: date-time
        hrCreator:
          $ref: "#/components/schemas/UserDto"
      required:
        - jobID
        - title
        - status
        - hrCreator
...
    ApplicationDto:
      type: object
      properties:
        applicationID:
          type: string
          format: uuid
        submissionTimestamp:
          type: string
          format: date-time
        status:
          type: string
          enum: [
              SUBMITTED, # Submitted
              AI_SCREENING, # AI screening
              AI_INTERVIEW, # AI interview
              COMPLETED, # Assessment completed
              SHORTLISTED, # Shortlisted
              REJECTED, # Rejected
              HIRED, # Hired
            ]
        resumeContent:
          type: string
          description: Resume text content
        originalResumeFilename:
          type: string
          description: Original resume filename
        lastModifiedTimestamp:
          type: string
          format: date-time
        candidate:
          $ref: "#/components/schemas/UserDto"
        job:
          $ref: "#/components/schemas/JobDto"
        assessment:
          $ref: "#/components/schemas/AssessmentDto"
      required:
        - applicationID
        - status
        - candidate
        - job
    CreateApplicationRequest:
      type: object
      properties:
        jobID:
          type: string
          format: uuid
        resumeContent:
          type: string
          description: Resume text content
        originalResumeFilename:
          type: string
          description: Original resume filename (optional)
      required:
        - jobID
        - resumeContent

    # AI assessment-related models
    AssessmentDto:
      type: object
      properties:
        assessmentID:
          type: string
          format: uuid
        applicationID:
          type: string
          format: uuid
        resumeScore:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Resume AI score (0-100)
        interviewScore:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Interview AI score (0-100)
        finalScore:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Final composite score (0-100)
        resumeAnalysis:
          type: string
          description: Resume analysis results
        interviewSummary:
          type: string
          description: Interview summary
        recommendation:
          type: string
          enum: [RECOMMEND, CONSIDER, NOT_RECOMMEND]
          description: AI recommendation result
        creationTimestamp:
          type: string
          format: date-time
        lastModifiedTimestamp:
          type: string
          format: date-time
      required:
        - assessmentID
        - applicationID

    # AI chat-related models
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags:
        - Authentication
      summary: User login
      description: User (HR or Candidate) login using email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: "john@example.com"
              password: "password123"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags:
        - Authentication
      summary: Candidate registration
      description: New candidate account registration, defaults to CANDIDATE role
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              fullName: "John Smith"
              email: "john.smith@example.com"
              password: "password123"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags:
        - Authentication
      summary: User logout
      description: User logout and invalidate token
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /auth/profile:
    get:
      operationId: getProfile
      tags:
        - Authentication
      summary: Get current user information
      description: Get detailed information of the current logged-in user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Get successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserDto"

  # HR user management
...
  /applications:
    get:
      operationId: getApplicationList
      tags:
        - Application Management
      summary: Get current user's applications
      description: Candidates view their own application history, HR views all applications
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: jobId
          in: query
          description: Filter by job ID (HR only)
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum:
              [
                SUBMITTED,
                AI_SCREENING,
                AI_INTERVIEW,
                COMPLETED,
                SHORTLISTED,
                REJECTED,
                HIRED,
              ]
      responses:
        "200":
          description: Get successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          content:
                            type: array
                            items:
                              $ref: "#/components/schemas/ApplicationDto"
                          pageInfo:
                            $ref: "#/components/schemas/PageInfo"

    post:
      operationId: createApplication
      tags:
        - Application Management
      summary: Submit job application
      description: Candidates apply for specified job positions and submit resumes
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
            example:
              jobID: "123e4567-e89b-12d3-a456-426614174000"
              resumeContent: "Name: John Smith\nEducation: Bachelor's in Computer Science\nWork Experience: 3 years Java development experience..."
              originalResumeFilename: "john_smith_resume.pdf"
      responses:
        "201":
          description: Application submitted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ApplicationDto"
        "400":
          description: Application failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions (non-Candidate user)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applications/{applicationId}:
    get:
      operationId: getApplicationDetails
      tags:
        - Application Management
      summary: Get application details
      description: Get detailed information and assessment results of a specific application
      security:
        - BearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Get successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ApplicationDto"
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ===================
  # AI Chat Interview APIs (GenAI Service)
  # ===================
...
security:
  - BearerAuth: []